+++ affinda-api-client/src/main/java/com/affinda/api/client/models/RowAnnotation.java
@@ -1,12 +1,7 @@
 package com.affinda.api.client.models;
 
 import com.azure.core.annotation.Fluent;
-import com.fasterxml.jackson.annotation.JsonAnyGetter;
-import com.fasterxml.jackson.annotation.JsonAnySetter;
-import com.fasterxml.jackson.annotation.JsonIgnore;
 import com.fasterxml.jackson.annotation.JsonProperty;
-import java.util.HashMap;
-import java.util.Map;
 
 /** The RowAnnotation model. */
 @Fluent
@@ -16,11 +11,6 @@ public final class RowAnnotation extends Annotation {
      */
     @JsonProperty(value = "parsed")
     private RowAnnotationParsed parsed;
-
-    /*
-     * Dictionary of <any>
-     */
-    @JsonIgnore private Map<String, Object> additionalProperties;
 
     /**
      * Get the parsed property: The parsed property.
@@ -40,34 +30,5 @@ public final class RowAnnotation extends Annotation {
     public RowAnnotation setParsed(RowAnnotationParsed parsed) {
         this.parsed = parsed;
         return this;
-    }
-
-    /**
-     * Get the additionalProperties property: Dictionary of &lt;any&gt;.
-     *
-     * @return the additionalProperties value.
-     */
-    @JsonAnyGetter
-    public Map<String, Object> getAdditionalProperties() {
-        return this.additionalProperties;
-    }
-
-    /**
-     * Set the additionalProperties property: Dictionary of &lt;any&gt;.
-     *
-     * @param additionalProperties the additionalProperties value to set.
-     * @return the RowAnnotation object itself.
-     */
-    public RowAnnotation setAdditionalProperties(Map<String, Object> additionalProperties) {
-        this.additionalProperties = additionalProperties;
-        return this;
-    }
-
-    @JsonAnySetter
-    void setAdditionalProperties(String key, Object value) {
-        if (additionalProperties == null) {
-            additionalProperties = new HashMap<>();
-        }
-        additionalProperties.put(key, value);
     }
 }