+++ CHANGELOG.md
@@ -12,6 +12,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Custom fields to Job Descriptions
 - Add custom data to job description search results
 - Add international_country_code to phone number details
+- Endpoints for add/remove tag for documents
 
 ### Changed
 - Provide additional filters for data point choices, and allow data point choices to be specified for any existing text field.